refactor(evaluate): extract shared DimensionSelector, reuse in EvaluationForm and ReEvaluateCard
